Created file for Email Check
This was a file that I was working on, didn't get to see if it worked, but its a start. Just placing it here in case you want to try and make this work.

onclick of the send request button was supposed to check the email on this page and then redirect to the index page. you can use formaction = "createAcctEmil.php" inside of the input field for the send request button.

Not sure this is the correct way to g, but its an option.